pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "step-stone"
version = "0.1.0"
description = "A project for automatic complex task breakdown using LLMs."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.47.0",
    "gradio>=5.49.1",
    "pymongo>=4.15.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.2.1",
]

[tool.setuptools.packages.find]
where = ["src"]
